core: add support for "weak" gapless audio and make it default
The main issue with gapless audio support is that it forces resampling
to the sample rate of the first file (or more precisely: the samplerate
the AO was opened - the AO can't be reconfigured). When playing multiple
files, and the first file had a low samplerate, this sample rate was
kept until the AO was closed.

Introduce a "weak" mode for gapless audio playback that tries to support
gapless playback only when advancing to files with the same samplerate.
If the samplerate changes, the AO is reopened (i.e. it can create gaps
in the speaker audio output). Make this mode default.

Heuristic that decides when to do gapless playback (i.e. no AO
reinitialization forced):
- the decoded sample rate matches the samplerate of the open AO
- the decoded sample rate is the same as the previous file
  (reduces number of useless reinitializations when the AO supports
   only a single samplerate)

Issues:
- this will block to drain the remaining audio if the new file is found
  to have a different samplerate
- user-inserted filters or forcing sample with --srate will interfere
  with the heuristic and probably destroy gapless playback
- maybe it should try to reopen the AO at a later point to deal with the
  issues mentioned above?
- it's not completely clear what happens in the ordered chapter case
- what about sample size and number of channels? Checking samplerate is
  only half the job, and this is probably critically missing.

The old behavior can be restored with --gapless-audio=no. When using
--gapless-audio, behavior should be the same as before the change.

Note: due to shortcomings of the option parser, --no-gapless-audio won't work anymore.