stream: report chapter times
Note that the cache could make this excessively slow, because the
frontend asks for chapter times very often (e.g. all the time when the
OSD bar is visible). Possibly the chapters should be cached instead,
but then we'd need to detect when the chapters change.

As of now stream_dvd uses some sort of quadratic algorithm to get the
time for a chapter, but hopefully this doesn't matter.